[Backend] Update scf.if result uses in RewriteTensorPointer pass (#4893)
I ran into an error in the RewriteTensorPointer pass. In my IR, there's
an scf.if that produces a non-pointer result. The rewriteIfOp() created
a new scf.if, but the use of scf.if result is still referencing the old one,
which caused a compile error. In this patch, I updated all uses of scf.if
with the results of the new if-op.