[BACKEND][NVIDIA] pass ptx-version to ttgir->llir conversion pass and use it for vectorized atomics (#4982)
Vectorized atomics on NVIDIA
(https://github.com/triton-lang/triton/pull/4971) are only available on
Hopper (>=sm90) and PTX >= 8.1. It's possible to be running with PTX 8.0
on a Hopper machine. This PR passes ptx-version to the ttgir->llir
conversion pass for NVIDIA, and uses the ptx version to determine
whether vectorized atomics should be used.